Fixed degrees to radians op

+++ scrimage-core/src/main/java/com/sksamuel/scrimage/angles/Degrees.java
@@ -8,6 +8,6 @@ public class Degrees {
    }
 
    public Radians toRadians() {
-      return new Radians((int) ((double) value * Math.PI / 180.0));
+      return new Radians(((double) value * Math.PI / 180.0));
    }
 }
